move processor classes to processor directory
